fix sanic type issues for good
