--- stubs/sanic/__init__.pyi
@@ -1,18 +1,0 @@
-from sanic.__version__ import __version__
-from sanic.app import Sanic
-from sanic.blueprints import Blueprint
-from sanic.constants import HTTPMethod
-from sanic.request import Request
-from sanic.response import HTTPResponse, html, json, text
-
-__all__ = [
-    "__version__",
-    "Sanic",
-    "Blueprint",
-    "HTTPMethod",
-    "HTTPResponse",
-    "Request",
-    "html",
-    "json",
-    "text",
-]
