--- stubs/sanic/mixins/exceptions.pyi
@@ -1,8 +1,0 @@
-from sanic.mixins.routes import RouteWrapper
-from typing_extensions import Protocol
-
-
-class ExceptionMixin(Protocol):
-
-    def exception(self, *exceptions: Exception, apply: bool =True) -> RouteWrapper:
-        ...
